negative amount issue fixed and applyed coupon should not get removed automatically

+++ dbcenterapi/src/main/java/com/boost/dbcenterapi/utils/DataLoader.kt
@@ -175,7 +175,7 @@ object DataLoader {
                                     coupons.add(
                                         CouponsModel(
                                             singleCoupon.code,
-                                            singleCoupon.discount_percent
+                                            singleCoupon.discount_percent.toDouble()
                                         )
                                     )
                                 }
