+++ marketplace/src/main/java/com/boost/marketplace/ui/home/MarketPlaceHomeViewModel.kt
@@ -377,7 +377,7 @@ class MarketPlaceHomeViewModel() : BaseViewModel() {
                                 coupons.add(
                                     CouponsModel(
                                         singleCoupon.code,
-                                        singleCoupon.discount_percent
+                                        singleCoupon.discount_percent.toDouble()
                                     )
                                 )
                             }
